move ScreenLockProvider to Store
