+++ common/v2/features/ScreenLock/ScreenLockNew.tsx
@@ -5,8 +5,8 @@ import styled from 'styled-components';
 import translate, { translateRaw } from 'translations';
 
 import { ExtendedContentPanel, InputField } from 'v2/components';
-import { ScreenLockContext } from 'v2/providers';
 import { AnalyticsService, ANALYTICS_CATEGORIES } from 'v2/services';
+import { ScreenLockContext } from './ScreenLockProvider';
 
 // Legacy
 import mainImage from 'common/assets/images/icn-create-pw.svg';
