+++ common/v2/features/ScreenLock/index.ts
@@ -2,3 +2,4 @@ export { default as ScreenLockLocking } from './ScreenLockLocking';
 export { default as ScreenLockNew } from './ScreenLockNew';
 export { default as ScreenLockLocked } from './ScreenLockLocked';
 export { default as ScreenLockForgotPassword } from './ScreenLockForgotPassword';
+export { ScreenLockContext, default as ScreenLockProvider } from './ScreenLockProvider';
