+++ common/v2/features/index.ts
@@ -1,5 +1,4 @@
 export * from './Wallets';
-export { ScreenLockLocking } from './ScreenLock';
 export { Drawer, AccountScreen, UnlockScreen, SelectLanguage } from './Drawer';
 export { DevTools } from './DevTools';
 export { Home } from './Home';
@@ -9,7 +8,14 @@ export { AddAccountFlow } from './AddAccount';
 export { Settings, Import, Export } from './Settings';
 export { DownloadApp } from './DownloadApp';
 export { NoAccounts } from './NoAccounts';
-export { ScreenLockNew, ScreenLockLocked, ScreenLockForgotPassword } from './ScreenLock';
+export {
+  ScreenLockNew,
+  ScreenLockLocked,
+  ScreenLockForgotPassword,
+  ScreenLockLocking,
+  ScreenLockProvider,
+  ScreenLockContext
+} from './ScreenLock';
 export { SendAssets } from './SendAssets';
 export { ReceiveAssets } from './ReceiveAssets';
 export { BuyAndExchange, ShapeShiftAuthorization, ZeroEx } from './BuyAndExchange';
