--- common/v2/providers/ScreenLockProvider/index.ts
@@ -1,1 +1,0 @@
-export { ScreenLockContext, default as ScreenLockProvider } from './ScreenLockProvider';
